update api get/update/post user and userprofile

+++ src/main/java/com/java/web_travel/repository/UserProfileRepository.java
@@ -3,5 +3,5 @@ package com.java.web_travel.repository;
 import com.java.web_travel.entity.UserProfileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserProfileRepository extends JpaRepository<UserProfileEntity, Long> {
+public interface UserProfileRepository extends JpaRepository<UserProfileEntity, Integer> {
 }
