+++ src/main/java/com/java/web_travel/repository/UserRepository.java
@@ -11,8 +11,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     Optional<UserEntity> findByTelephone(String telephone); // Add this method
     // Other methods...
     UserEntity findByUserName(String username);
+
 }
